requirements.txt
django
django-import-export
django-cors-headers
djangorestframework
pillow 